Refactor backend
