+++ backend/src/database/addSimpleDataToCollection.ts
