+++ backend/src/database/addSpotifyDataToCollection.ts
@@ -1,8 +1,8 @@
 require('dotenv').config({ path: '../config.env' });
+import { get } from '../http/request';
 import _ from 'lodash';
 import { getSpotifyAuth } from '../helpers/getSpotifyAuth';
 import { removeDuplicateArtists } from '../helpers/removeDuplicateArtists';
-import { get } from '../http/request';
 const stringSimilarity = require('string-similarity');
 
 export const updateCollectionWithSpotify = async (collection_name: string, db_connect: any) => {
@@ -39,34 +39,23 @@ async function addSpotifyMainData(element: any, token: any) {
         type: 'artist',
       },
     })
-      .then(async function (res) {
-        //console.log('res', res);
+      .then(async function (res: any) {
         const bestMatch = findBestMatch(
           res.data.artists.items,
           element.artist,
           res.data.artists.items.map((artist: any) => artist.name),
         );
 
-        //console.log('bestMatch', bestMatch);
         try {
           element.band_id = bestMatch.id;
           element.sp_band_name = bestMatch.name;
           element.link = bestMatch.external_urls.spotify;
           element.uri = bestMatch.uri;
-          //element.genres = res.data.artists.items[0].genres;
         } catch {}
-
-        // try {
-        //   element.band_id = res.data.artists.items[0].id;
-        //   element.sp_band_name = res.data.artists.items[0].name;
-        //   element.link = res.data.artists.items[0].external_urls.spotify;
-        //   element.uri = res.data.artists.items[0].uri;
-        //   //element.genres = res.data.artists.items[0].genres;
-        // } catch {}
 
         resolve();
       })
-      .catch(function (error) {
+      .catch(function (error: any) {
         console.log('Error: addSpotifyMainData');
         console.log(error.response);
       });
@@ -80,7 +69,6 @@ function findBestMatch(artists: any, query: any, options: any) {
   );
 
   if (matches.bestMatch.rating >= 0.95) {
-    //return options[matches.bestMatchIndex];
     return artists[matches.bestMatchIndex];
   }
   return null;
@@ -95,7 +83,7 @@ async function addSpotifyTopTracks(element: any, token: any) {
           Authorization: 'Bearer ' + token,
         },
       })
-        .then(async function (res) {
+        .then(async function (res: any) {
           try {
             element.albumArtUrl = res.data.tracks[0].album.images[1].url;
             element.topTrackURIs = res.data.tracks.map((track: any) => track.uri);
@@ -106,7 +94,7 @@ async function addSpotifyTopTracks(element: any, token: any) {
 
           resolve();
         })
-        .catch(function (error) {
+        .catch(function (error: any) {
           console.log(error);
         });
     });
