+++ backend/src/enums/Cities.ts
@@ -1,0 +1,4 @@
+export enum Cities {
+  Victoria = 'victoria',
+  Vancouver = 'vancouver',
+}
