+++ backend/src/enums/Festivals.ts
@@ -1,8 +1,3 @@
-export enum Cities {
-  Victoria = 'victoria',
-  Vancouver = 'vancouver',
-}
-
 export enum Festivals {
   PhilipsBackyard = 'phillipsBackyard',
   Rifflandia = 'rifflandia',
