+++ backend/src/helpers/createPlaylist.ts
@@ -1,4 +1,3 @@
-import { Cities, Festivals } from '../enums/common';
 import { SpotifyPlaylistDataType } from '../types/SpotifyTypes';
 import { PLAYLIST_IMG_RS } from '../assets/recordshop_img';
 import { sortByPopularity } from './sortByPopularity';
@@ -6,6 +5,8 @@ import { sortByDateAndOrder } from './sortByDateAndOrder';
 import { filterRecent } from './filterRecent';
 import { get, post, put } from '../http/request';
 import { HttpRequestError } from '../http/HttpRequestError';
+import { Cities } from '../enums/Cities';
+import { Festivals } from '../enums/Festivals';
 
 export const CreateNewPlaylistJamBase = async ({
   token,
