+++ backend/src/index.ts
@@ -1,17 +1,24 @@
-import healthchecker from './routes/healthchecker';
 import express from 'express';
 import cors from 'cors';
+
 require('dotenv').config({ path: '../.env' });
 const port = process.env.PORT || 5000;
-import { recordRoutes } from './routes/record';
 const app = express();
-// get driver connection
-import dbo from './db/conn';
+
+import dbo from './database/conn';
+
+import healthchecker from './controller/healthchecker';
+import { victoriaRouter } from './controller/victoriaController';
+import { rifflandiaRouter } from './controller/rifflandiaController';
+import { jamBaseRouter } from './controller/jamBaseController';
 
 app.use(cors());
 app.use(express.json());
-app.use(recordRoutes);
 app.use('/', healthchecker);
+
+app.use(victoriaRouter);
+app.use(rifflandiaRouter);
+app.use(jamBaseRouter);
 
 app.listen(port, async () => {
   // perform a database connection when server starts
