--- backend/src/rifflandia/constants.ts
@@ -1,2 +1,0 @@
-export const RIFFLANDIA_SIMPLE = 'rifflandia_simple';
-export const RIFFLANDIA_SPOTIFY = 'rifflandia';
