+++ backend/src/types/Artist.ts
@@ -1,0 +1,14 @@
+// TODO: This is the ideal data model, should use this instead of Artists.ts
+export interface ArtistNew {
+  name: string;
+  date: string;
+  venue: string;
+  popularity?: number;
+
+  spotify: {
+    albumArtUrl: string;
+    artistId: string;
+    link: string;
+    topTrackURIs: string[];
+  };
+}
